WorldService tests

+++ src/main/java/com/sharkio/backend/repository/PlayerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.sharkio.backend.model.Player;
 
 @Repository
-public interface PlayerRepository extends CrudRepository<Player, Long>{
+public interface PlayerRepository extends CrudRepository<Player, Integer>{
 }
